KS-21 feat: legacy entity, repository 이동
