+++ src/main/java/com/kreamish/kream/comment/entity/Comment.java
@@ -2,7 +2,7 @@ package com.kreamish.kream.comment.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import com.kreamish.kream.item.entity.Item;
-import com.kreamish.kream.legacy.entity.Member;
+import com.kreamish.kream.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
