+++ src/main/java/com/kreamish/kream/member/entity/Member.java
@@ -1,4 +1,4 @@
-package com.kreamish.kream.legacy.entity;
+package com.kreamish.kream.member.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import jakarta.persistence.Column;
@@ -12,27 +12,28 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.validator.constraints.Length;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "trade")
-public class Trade extends BaseEntity {
+@Table(name = "member")
+public class Member extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "trade_id")
-    private Long tradeId;
+    @Column(name = "member_id")
+    private Long memberId;
 
     @ManyToOne
-    @JoinColumn(name = "purchase_id")
-    private Purchase purchase;
+    @JoinColumn(name = "member_role_id")
+    private MemberRole memberRole;
 
-    @ManyToOne
-    @JoinColumn(name = "sale_id")
-    private Sale sale;
+    @Column(nullable = false)
+    @Length(max = 200)
+    private String email;
 
-    @Column(name = "trade_price", nullable = false)
-    private Long tradePrice;
-
+    @Column(nullable = false)
+    @Length(max = 256)
+    private String password;
 }
