+++ src/main/java/com/kreamish/kream/purchase/entity/Purchase.java
@@ -1,7 +1,9 @@
-package com.kreamish.kream.legacy.entity;
+package com.kreamish.kream.purchase.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import com.kreamish.kream.itemsizes.entity.ItemSizes;
+import com.kreamish.kream.common.entity.DealStatus;
+import com.kreamish.kream.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -13,23 +15,19 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "sale")
-@Builder
-@AllArgsConstructor
-public class Sale extends BaseEntity {
+@Table(name = "purchase")
+public class Purchase extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "sale_id")
-    private Long saleId;
+    @Column(name = "purchase_id")
+    private Long purchaseId;
 
     @ManyToOne
     @JoinColumn(name = "item_sizes_id")
@@ -39,10 +37,10 @@ public class Sale extends BaseEntity {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @Column(nullable = false)
-    private Long salePrice;
+    @Column(name = "purchase_price", nullable = false)
+    private Long purchasePrice;
 
     @Enumerated(EnumType.STRING)
-    @Column(name = "sale_status", nullable = false)
-    private DealStatus saleStatus;
+    @Column(name = "purchase_status", nullable = false)
+    private DealStatus purchaseStatus;
 }
