+++ src/main/java/com/kreamish/kream/rel/repository/ItemCollectionRelRepository.java
@@ -1,6 +1,6 @@
-package com.kreamish.kream.legacy.repository;
+package com.kreamish.kream.rel.repository;
 
-import com.kreamish.kream.legacy.entity.ItemCollectionRel;
+import com.kreamish.kream.rel.entity.ItemCollectionRel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemCollectionRelRepository extends JpaRepository<ItemCollectionRel, Long> {
