+++ src/main/java/com/kreamish/kream/sale/entity/Sale.java
@@ -1,7 +1,9 @@
-package com.kreamish.kream.legacy.entity;
+package com.kreamish.kream.sale.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import com.kreamish.kream.itemsizes.entity.ItemSizes;
+import com.kreamish.kream.common.entity.DealStatus;
+import com.kreamish.kream.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -13,19 +15,23 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "purchase")
-public class Purchase extends BaseEntity {
+@Table(name = "sale")
+@Builder
+@AllArgsConstructor
+public class Sale extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "purchase_id")
-    private Long purchaseId;
+    @Column(name = "sale_id")
+    private Long saleId;
 
     @ManyToOne
     @JoinColumn(name = "item_sizes_id")
@@ -35,10 +41,10 @@ public class Purchase extends BaseEntity {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @Column(name = "purchase_price", nullable = false)
-    private Long purchasePrice;
+    @Column(nullable = false)
+    private Long salePrice;
 
     @Enumerated(EnumType.STRING)
-    @Column(name = "purchase_status", nullable = false)
-    private DealStatus purchaseStatus;
+    @Column(name = "sale_status", nullable = false)
+    private DealStatus saleStatus;
 }
