+++ src/main/java/com/kreamish/kream/trade/entity/Trade.java
@@ -1,6 +1,8 @@
-package com.kreamish.kream.legacy.entity;
+package com.kreamish.kream.trade.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
+import com.kreamish.kream.purchase.entity.Purchase;
+import com.kreamish.kream.sale.entity.Sale;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -12,28 +14,27 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "member")
-public class Member extends BaseEntity {
+@Table(name = "trade")
+public class Trade extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
-    private Long memberId;
+    @Column(name = "trade_id")
+    private Long tradeId;
 
     @ManyToOne
-    @JoinColumn(name = "member_role_id")
-    private MemberRole memberRole;
+    @JoinColumn(name = "purchase_id")
+    private Purchase purchase;
 
-    @Column(nullable = false)
-    @Length(max = 200)
-    private String email;
+    @ManyToOne
+    @JoinColumn(name = "sale_id")
+    private Sale sale;
 
-    @Column(nullable = false)
-    @Length(max = 256)
-    private String password;
+    @Column(name = "trade_price", nullable = false)
+    private Long tradePrice;
+
 }
