+++ src/test/java/com/kreamish/kream/item/service/ItemServiceTest.java
@@ -9,8 +9,8 @@ import com.kreamish.kream.item.dto.ItemListResponseDto;
 import com.kreamish.kream.item.dto.ItemListSearchCondition;
 import com.kreamish.kream.item.entity.Item;
 import com.kreamish.kream.itemsizes.entity.ItemSizes;
-import com.kreamish.kream.legacy.entity.DealStatus;
-import com.kreamish.kream.legacy.entity.Sale;
+import com.kreamish.kream.common.entity.DealStatus;
+import com.kreamish.kream.sale.entity.Sale;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.util.ArrayList;
